pyproject.toml
[project]
name = "agent-team"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
 dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "black>=24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

testpaths = ["tests"]

[tool.black]
line-length = 100

target-version = ["py311"]

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
